src/main/java/com/minsait/junit/models/Cuenta.java
package com.minsait.junit.models;

import com.minsait.junit.exceptions.DineroInsuficienteException;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.Setter;

import java.math.BigDecimal;

@Getter
@Setter
@EqualsAndHashCode
public class Cuenta {
    private String persona;
    private BigDecimal saldo;
    private Banco banco;

    public Cuenta(String persona, BigDecimal saldo) {
        this.persona = persona;
        this.saldo = saldo;
    }

    public void retirar(BigDecimal monto) {
        BigDecimal saldoAux = this.saldo.subtract(monto);
        if(saldoAux.compareTo(BigDecimal.ZERO) < 0){
            throw new DineroInsuficienteException("Dinero Insuficiente");
        }
        this.saldo = saldoAux;
    }

    public void depositar(BigDecimal monto){
        this.saldo = this.saldo.add(monto);
    }
}
